+++ packages/next/src/lib/api.ts
@@ -1,16 +1,20 @@
 import { redirect } from "next/navigation"
 import { cookies, headers } from "next/headers"
 import type { AuthInstance, Session, User } from "@aura-stack/react"
-import type { NextSignInCredentials, NextSignInReturn, NextSignOutReturn } from "@/@types/api"
+import type {
+    NextSignInCredentials,
+    NextSignInReturn,
+    NextSignOutReturn,
+    NextUpdateSessionOptions,
+    NextUpdateSessionReturn,
+} from "@/@types/api"
 import type {
     GetSessionAPIOptions,
     SignInAPIOptions,
     SignOutAPIOptions,
-    DeepPartial,
     LiteralUnion,
     BuiltInOAuthProvider,
     SignInCredentialsAPIOptions,
-    UpdateSessionAPIReturn,
 } from "@aura-stack/react/types"
 
 /**
@@ -57,14 +61,17 @@ export const signIn = <DefaultUser extends User = User>({ api }: AuthInstance<De
         if (options?.redirect === false) {
             return signIn as NextSignInReturn<Options>
         }
-        return redirect(signIn.signInURL) as NextSignInReturn<Options>
+        if (signIn.success && options?.redirectTo) {
+            return redirect(signIn.signInURL)
+        }
+        return signIn as NextSignInReturn<Options>
     }
 }
 
 export const signInCredentials = <DefaultUser extends User = User>({ api }: AuthInstance<DefaultUser>) => {
-    return async <O extends SignInCredentialsAPIOptions>(
+    return async <Options extends SignInCredentialsAPIOptions>(
         options: SignInCredentialsAPIOptions
-    ): Promise<NextSignInCredentials<O>> => {
+    ): Promise<NextSignInCredentials<Options>> => {
         const signIn = await api.signInCredentials({
             headers: await headers(),
             ...options,
@@ -72,25 +79,26 @@ export const signInCredentials = <DefaultUser extends User = User>({ api }: Auth
         })
         await applyCookies(signIn.headers)
         if (signIn.success && options?.redirectTo) {
-            redirect(signIn.redirectURL)
+            return redirect(signIn.redirectURL)
         }
-        return signIn as NextSignInCredentials<O>
+        return signIn as NextSignInCredentials<Options>
     }
 }
 
 export const updateSession = <DefaultUser extends User = User>({ api }: AuthInstance<DefaultUser>) => {
-    return async (
-        session: DeepPartial<Session<DefaultUser>>,
-        options?: GetSessionAPIOptions
-    ): Promise<UpdateSessionAPIReturn<DefaultUser>> => {
+    return async <Options extends NextUpdateSessionOptions<DefaultUser>>(
+        options: Options
+    ): Promise<NextUpdateSessionReturn<Options, DefaultUser>> => {
         const updated = await api.updateSession({
-            session,
-            headers: await headers(),
-            skipCSRFCheck: true,
             ...options,
+            session: options.session,
+            headers: await headers(),
         })
         await applyCookies(updated.headers)
-        return updated
+        if (updated.success && options?.redirectTo) {
+            return redirect(updated.redirectURL)
+        }
+        return updated as NextUpdateSessionReturn<Options, DefaultUser>
     }
 }
 
@@ -110,10 +118,87 @@ export const signOut = <DefaultUser extends User = User>({ api }: AuthInstance<D
 
 export const api = <DefaultUser extends User = User>(config: AuthInstance<DefaultUser>) => {
     return {
+        /**
+         * Retrieves the current session data from the server-side.
+         *
+         * @param options - Options for the API call, including headers to verify `session_token` cookie.
+         * @returns An object containing session data see {@link User}
+         */
         getSession: getSession<DefaultUser>(config),
+        /**
+         * Initiates the sign-in flow on the server-side. By default the redirect is automatic, but it can be
+         * disabled by setting the `redirect` option to `false`. When redirect is disabled, the API returns the
+         * `signInURL` in the response for the client to handle the redirect manually.
+         *
+         * @param oauth - The OAuth provider to use for sign-in (e.g., "github", "gitlab", "bitbucket").
+         * @param options - Optional parameters for the sign-in process, including headers and redirect behavior.
+         * @returns The object returned by the API call {@link SignInAPIReturn}
+         * @example
+         * import { headers } from "next/headers"
+         *
+         * const response = await api.signIn("github", {
+         *   redirectTo: "/dashboard",
+         *   headers: await headers()
+         * })
+         */
         signIn: signIn<DefaultUser>(config),
+        /**
+         * Signs in a user using credentials (`username` and `password`) on the server-side. The credentials must
+         * be verified by the `authorize` function provided in the `credentials` configuration option.
+         *
+         * @param options - Options for the API call, including the credentials payload, headers, and redirect behavior.
+         * @returns The object returned by the API call {@link SignInCredentialsAPIReturn}
+         * @example
+         * import { headers } from "next/headers"
+         *
+         * const response = await api.signInCredentials({
+         *   payload: {
+         *     username: "johndoe",
+         *     password: "1234567890"
+         *   },
+         *   redirectTo: "/dashboard",
+         *   headers: await headers()
+         * })
+         */
         signInCredentials: signInCredentials<DefaultUser>(config),
+        /**
+         * Updates the current session on the server-side. It allows partial updates to the session object, such as
+         * modifying user fields or extending the session expiry. It implements CSRF Protection by default, for
+         * server-side calls it only verifies and validates the CSRF Token, it also provides Double-Submit
+         * Cookie protection by requiring the `session_token` cookie to be included in the request headers.
+         *
+         * @param options - Options for the API call, including the session updates, headers, redirect behavior, and CSRF check bypass.
+         * @returns The object returned by the API call {@link UpdateSessionAPIReturn}
+         * @example
+         * import { headers } from "next/headers"
+         *
+         * const response = await api.updateSession({
+         *   session: {
+         *     user: {
+         *       name: "John Doe",
+         *       email: "john.doe@example.com"
+         *     }
+         *   },
+         *   redirectTo: "/dashboard",
+         *   headers: await headers()
+         * })
+         */
         updateSession: updateSession<DefaultUser>(config),
+        /**
+         * Signs out the current session on the server-side. It implements CSRF Protection by default, for
+         * server-side calls it only verifies and validates the CSRF Token, it also provides Double-Submit
+         * Cookie protection by requiring the `session_token` cookie to be included in the request headers.
+         *
+         * @param options - Options for the API call, including headers, redirect behavior, and CSRF check bypass.
+         * @returns The object returned by the API call {@link SignOutAPIReturn}
+         * @example
+         * import { headers } from "next/headers"
+         *
+         * const response = await api.signOut({
+         *   redirectTo: "/goodbye",
+         *   headers: await headers(),
+         * })
+         */
         signOut: signOut<DefaultUser>(config),
     }
 }